✅ Chore: 파일 위치 변경 및 필요없는 파일 삭제 (#125)
