+++ NadoSunbae-iOS/NadoSunbae-iOS/Network/APIModels/Models/Classroom/InfoDetail/InfoDetailDataModel.swift
@@ -1,0 +1,8 @@
+//
+//  InfoDetailDataModel.swift
+//  NadoSunbae-iOS
+//
+//  Created by hwangJi on 2022/02/15.
+//
+
+import Foundation
