+++ NadoSunbae-iOS/NadoSunbae-iOS/Network/APIModels/Models/Classroom/Question/ClassroomQuestionDetailData.swift
@@ -62,3 +62,4 @@ func convertToMajorInfoString(_ firstMajorName: String, _ firstMajorStart: Strin
         return firstMajorName + " " + firstMajorStart + " " + "|" + " " + secondMajorName + " " + secondMajorStart
     }
 }
+
