+++ NadoSunbae-iOS/NadoSunbae-iOS/Network/APIModels/Models/Public/PostLikeResModel.swift
@@ -1,0 +1,8 @@
+//
+//  PostLikeResModel.swift
+//  NadoSunbae-iOS
+//
+//  Created by hwangJi on 2022/02/15.
+//
+
+import Foundation
